+++ WebApplication1/Models/Database/GroupAdmins.cs
@@ -6,7 +6,7 @@ namespace WebApplication1.Models.Database
     public class GroupAdmins : BaseEntity
     {
         [ForeignKey(nameof(Group))]
-        public int IdGroup { get; set; }
+        public long IdGroup { get; set; }
         public virtual Groups Group { get; set; }
 
         [ForeignKey(nameof(User))]
