+++ WebApplication1/Models/Database/Scenarios.cs
@@ -13,7 +13,7 @@ namespace WebApplication1.Models.Database
         public ScenarioActions Action { get; set; }
 
         [ForeignKey(nameof(Group))]
-        public int IdGroup { get; set; }
+        public long IdGroup { get; set; }
         public virtual Groups Group { get; set; }
 
         [ForeignKey(nameof(Message))]
