+++ WebApplication1/Models/Database/Segments.cs
@@ -8,8 +8,9 @@ namespace WebApplication1.Models.Database
         public string Name { get; set; }
         [Column(TypeName = "datetime2")]
         public DateTime DtCreate { get; set; }
+
         [ForeignKey(nameof(Group))]
-        public int IdGroup { get; set; }
+        public long IdGroup { get; set; }
         public virtual Groups Group { get; set; }
     }
 }
