+++ WebApplication1/Models/Database/Users.cs
@@ -4,7 +4,7 @@ namespace WebApplication1.Models.Database
 {
     public class Users : IdentityUser
     {
-        public int IdVk { get; set; }
-        public int IdCurrentGroup { get; set; }
+        public long IdVk { get; set; }
+        public long IdCurrentGroup { get; set; }
     }
 }
